Improve command output when nix fails
parse_setuppy_data.nix is expected to fail in some occasions. But
nix-build could fail in unexpected cases, for example when we run out of
disk space. This change tries to difference between the legit fails and
the unexpected ones. When an unexpected fail occurs, the command will
propagate nix-build's stderr.